refactor: Added primitives folder
