+++ src/core/api.h
@@ -8,7 +8,7 @@
 #include "rt3.h"
 #include "../integrators/integrator.h"
 #include "../shapes/sphere.h"
-#include "primlist.h"
+#include "../primitives/primlist.h"
 
 //=== API Macro definitions
