+++ src/core/rt3-base.h
@@ -9,12 +9,12 @@
 #include "image_io.h"
 #include "paramset.h"
 #include "parser.h"
-#include "primitive.h"
+#include "../primitives/primitive.h"
 #include "surfel.h"
 #include "../shapes/sphere.h"
 #include "../materials/material.h"
 #include "../shapes/shape.h"
-#include "primlist.h"
+#include "../primitives/primlist.h"
 #include "scene.h"
 #include "../integrators/integrator.h"
